rw_tryupgrade take 2
lck_rw_lock_shared_to_exclusive() on XNU will always block waiting on readers
to drain, which is not how rw_tryupgrade() is designed to work. (The "try" part).
So we rework it to use rw_tryenter() if readers == 1, and otherwise fail.

There is a worst case situation where the caller with block waiting on a READER
lock, in the failure branch.

Add additional ASSERTS to ensure correctness.