be a bit less gentle on available memory
Memory is given back just fine, but we may be forcing too much pressure
thus giving back a bit too much in the face of a brief period (milliseconds)
of pressure.

The internal fragmentation isn't very large, so use the amount of
system-allocated but not-in-low-arenas [bucket_heap's mem_inuse is the
most accessible manifestation of that] as a stronger signal.

Additionally, we should consider the amount of total memory not used
by spl if we are below the 70% threshold.  Inflating spl_free is
reasonably harmless, since the allocator interface is itself sensitive
to low memory.   We also have to do some inflating because we do not
have a fill thread speculatively increasing the overall size of vmem.