allocations and frees go to correct buckets
Doesn't make a functional difference, but we want a 2 MiB allocation to
be in bucket_2097152 and not bucket_4194304, since the bucket numbers
are defined to be the smallest possible allocation in that bucket.

Also, this would be functionally important if we end up adjusting the
16MiB allocation size per bucket (e.g. to use smaller chunks on
smaller-memory systems, either for all buckets or following the qcache
logic of highest power of two above 3 * [bucket_]size).