spin-sleep to avoid over-allocation.
Solve this problem:

1. bucket_heap is asked to satisfy an allocation and cannot,
   so it calls vmem_bucket_alloc()

2. meanwhile, bucket_heap is asked again to satisfy an allocation
   in the same size bracket, and again calls vmem_bucket_alloc().

3. xnu is asked for 2 x 16 MiB allocations.   both can succeed; one
   can succeed and the other no longer has enough memory and so waits
   in xnu_alloc_throttled(), but in so doing calls
   spl_free_set_emergency_pressure(); it may also bail from
   xnu_allocate_throttled() and trigger a kmem_reap().

Avoid this by serializing allocations to an individual bucket (using its
arena mutex), and by using a spin-sleep lock to break up concurrent
allocations to buckets that would have to call xnu_alloc_throttled().

This also reduces the chance of a spl_free_set_emergency_pressure() in
the face of multiple concurrent allocations to different buckets when
memory is nearly full.

Note that the races above can be wider than two-way races.   8-way races
have been observed in the wild!