128kiB bucket often wastes ~ 1/4 of its memory
Observation and analysis suggest that a step-down for that particular
bucket from 16 MiB to 4 MiB will avoid substantial waste caused by
concurrency-driven interleaved allocations from parent arenas that
cannot qcache at that size.    This leads to more "kidnapping" of
large spans in the bucket_131072 arena than any of the other arenas.

Shrinking that arena's minimum transfers to/from xnu by 1/4 will reduce
the amount of waste in SPL overall, and make it easier to shrink SPL
memory by smaller reductions in ARC size.

The 256 k arena can waste more memory in absolute terms, but because it
is the size that qcaching allocates, the percentage of memory wasted
is typically much smaller than 1/2, so a step size down will not be
worthwhile.

The 64k arena is bursty, but gives back spans extremely efficiently, so
it often wastes ~ 16 MiB and has a single-span allocation.   It is not
worth special-casing the 64k arena like the 128k arena because the
burstiness imposes more work on the xnu allocator and free-list management.