Be gentle in spl_free_thread() if we just allocated
Make sure spl_xat_lastalloc is uptime-in-seconds of the most
recent SUCCESSFUL alloc.

If we have successfully allocated in the last second, then avoid doing
an immediate free via pressure in spl_free_thread().

Since we are allocating in large pieces (16 MiB typically) we might not
be able to do two back-to-back allocs, and spl_free_thread() should not
freak out in that case.

This will slightly increase pressure on userland, which is worth it to
avoid an unnecessary shrink of ARC.

squash! Be gentle in spl_free_thread() if we just allocated