reduce default minimum span sizes in vmem_init()
better serialization means we are fine with smaller xnu spans

smaller spans -> less internal fragmentation

and as we do only power-of-two spans and serialize allocs
and frees, we perform well with pretty small spans.   bigger ones,
however, are more efficient; it's always better never to
talk to xnu's allocator at all, and the segment mechanism
in vmem is great (and better with big spans in the face of
nontrivial amounts of non power of two allocations from
above the bucket_heap arena, which in fact happens).