fix lundman's mutex reuse-after-mutex-destroy problem
The zio arenas have qcaches on them that have to be
destroyed before the kmem system is terminated, and before
the zio arenas themselves are destroyed.

reorder spl_kmem_fini()'s call segkmem_zio_fini() to just
after kmem_cache_fini().  This casues the vmem arenas to
be destroyed after their own qcaches, but before the
metadata arenas are destroyed.