avoid long pauses: waiters in vmem_xalloc() should be awakened
When the XAT or vmem_bucket alloc and free functions have
successfully finished talking to xnu, broadcast the news
to all waiters on the arena condvar; the waiters are in
the cv_wait of vmem_xalloc or vmem_nextalloc and unless
awakened by the allocators, they will be stuck for up to a
quarter minute until vmem_update() does its cv_broadcast().

This eliminates several-second stalls when doing huge
sequential writes.