more pressure and large pressure circumstances
If the arc is left to grow "overnight", it will do so, and
Activity Monitory will be yellow in the "morning", and
will take a bit of time to clear.   Although memory will
shrink fine on its own, the cosmetics of Activity Monitor
are a real consideration for O3X users.

Moreover, "overnight" when background processes do
significant I/O, there may be bursts where the bucket
arena fail kstats increase (these are failed VM_NOSLEEP
allocations, which are mostly driven by taskq and buf hash
manipulation in the zfs module).  Pressure after a first
failed attempt in XAT.

Finally, gently pressure if XAT_bail returns without
allocating; these will stil increment the wait kstat for
the arena, but should clear some room for when these
about-to-be-waited threads are re-awakened.