more pressure and large pressure circumstances
If the arc is left to grow "overnight", it will do so, and
Activity Monitory will be yellow in the "morning", and
will take a bit of time to clear.   Although memory will
shrink fine on its own, the cosmetics of Activity Monitor
are a real consideration for O3X users.

Moreover, "overnight" when background processes do
significant I/O, there were bursts where the bucket arena
fail kstats increase (these are failed VM_NOSLEEP
allocations, which are mostly driven by taskq and buf hash
manipulation in the zfs module).  Pressure after a first
failed attempt in XAT.  Although this has a small chance
of descheduling a VM_NOSLEEP thread because of a collision
with the mutex_enter(&spl_free_manual_pressure_lock), that
lock is not highly contended and XNU's underlying
lck_mtx_lock() is hybrid and spins briefly before
sleeping, and nothing holds the mutex long, so this does
not really break the VM_NOSLEEP does not wait for
resources description in the Illumos code.

The goal of the really large response is to roll back
long-term userland-idle / IO-non-idle
large-allocated-memory state when userland goes non-idle,
and should not happen regularly while a user is active on
the machine.

However we can also give a more muted (but still
relatively large) response for the secondary goal of
reducing the chance of back-to-back NULL returns from XAT
from slow growth from a moderate-IO, light-userland growth
of allocated memory.  This can fire more often.

Additionally, we gently pressure if XAT returns a NULL;
although this still results in taking an increment in the
.wait or .fail (depending on VM_SLEEP vs VM_NOSLEEP) kstat
for the arena, it does break up runs of them.

Penultimately, try a seocnd time to VM_NOSLEEP allocate in XAT
after pressure is sent; there is a small but non-zero
chance of a late allocation.

Finally, if we force a VM_PANIC alloc or get lucky with
the second VM_NOSLEEP attempt, update spl_xat_lastalloc,
and alert waiters appropriately.