temper the maybe large pressure reaction
The goal of the really large response is to roll back
long-term userland-idle / IO-non-idle
large-allocated-memory state when userland goes non-idle,
and should not happen regularly while a user is active on
the machine.

However we can also give a more muted response for the
secondary goal of reducing the chance of back-to-back
NULL returns from XAT from slow growth from a moderate-IO,
light-userland growth of allocated memory.   This can fire
more often.