pressure, then try a second time to VM_NOSLEEP allocate in XAT
VM_NOSLEEP allocations are allowed to fail by their
nature, and should not be delayed (they're usually done by
taskq or hashing trying to right-size resources), but it
only costs a few cycles to retry immediately after
pressure is flagged, even though it is extremely unlikely
to succeed.

Also when we force a VM_PANIC alloc or get lucky with the
second VM_NOSLEEP attempt, update spl_xat_lastalloc.