Improve handling of long vmem_xalloc() cv_wait queues
Iterate more through the XAT VM_SLEEP for loop when there
are waiters on the arena's cv_wait queue, or if there are
waiters on any arena's cv_wait queue.

Update "now" when we exit the cv_timedwait_hires().

Distinguish between late successful allocations to satisfy
the VM_NOSLEEP case and the VM_SLEEP case in XAT.

If there is a long arena cv_wait queue such that we force
in XAT_bail, update the exclusionary last_forced_ticks
variable.

Also kstat count the number of times we retain excess
allocations that previously would have been freed.

Finally, clarify a couple of comments in XAT.