make vmem_bucket_alloc() less of a bottleneck
We don't have to wait on the lower-level bucket vm_lock or
deal with its condvar, since vmem_alloc() and what it
calls will all do admission control.

We move quickly to vmem_alloc(lower-level-bucket, ...) if
there the lower-level bucket appears to be able to satisfy
the allocation without going to xnu, or if the allocation
vmflags disallow sleeping.   However, as we do a
vmem_canalloc(bmp, size) we must take care to preclude concurrent
vmem_alloc(bmp, size, ...) calls racing to the same
bucket when the bucket does not have enough memory for all
of them.   (Fortunately vmem_alloc() with INSTANTFIT is
very very fast, so bursts will be satisfied quickly).

Otherwise we spin-sleep for a few milliseconds waiting for
one of these conditions.   Eventually we simply call
vmem_alloc() and return its results.