contrived circumstances: mutex held for many seconds in vmem_xalloc()
If we make arc take a long time to actually shrink, then
the spl_free_set_and_wait_pressure() call can take a
while, thus holding the mutex for a long time, although
this generally takes fiddling with the arc size variables.

Drop and reacquire the mutex, as is practice for calls
that MAY take a long time in the vmem_alloc() and
vmem_xalloc().