don't mutex in normal case in xnu_free_throttled()
xnu_alloc_free_thread is usually uncontended, so there is
no need to take the lock in the normal case (i.e., when
a_waiters == 0 on entry), so we save on some function calls.

In the case where a burst of frees arrive, the for(;;)
loop will spread out the calls to xnu's free function.