let only one eligible bailer into XAT_bail at a time
Although xnu_allocate_throttled_bail() can handle
concurrent bailers, it runs a small risk of
over-allocating or needlessly bouncing a bailer back to
vmem_xalloc() with a NULL.  Neither is a bad outcome, but
serializing would-be bailers within XAT itself is more
efficient.

Gate with an _Atomic, which clang+llvm implements as
ncarefully arranged incw and decw instructions, which are
atomic because of natural alignment, but which do not
bus-lock.