A new spl memory system based on vmem.
Highlights:

* No more need to adjust arc_c_max or arc_c_min.
* Acitivity monitor's "Memory Pressure" graph should stay green.
* No glitches or out-of-memory conditions.
* Far less "wasted" memory.

Illumos's vmem/kmem systems are the native allocators of
their kernel, and the whole of the physical memory space
is divided up and given to low-level arenas that do vmem
metadata and the lowest-level heap, I/O space, and so forth.
kmem caching lives on top of the lowest-level vmem heap.

The kernel itself is vmem aware and generates free memory
variables that is used by arc and other users of the kmem.

This is a good and scalable system which we try to mimic
in the Solaris Porting Layer (SPL).  However, we are a
kernel extension and rely upon the macOS kernel (xnu) to
manage physical memory and generate variables suggesting
how much virtual memory is available to all clients,
including SPL.

Because we are just another client of XNU's memory
management system, and because the kernel API restricts,
we have to synthesize a variable, spl_free (in spl-kmem.c),
which is used by ARC to determine whether it should grow
or shrink.  ARC size is the principal factor in
determining the amount of memory we need from XNU.

arc.c (in the zfs kext) uses an external function call to
access spl_free.    arc.c also uses a small set of functions
to react swiftly to a pressure signal which is generated in
various parts of the spl module when XNU looks like it may
transition into memory relcamation.

We arrange a hierarchy of vmem arenas starting with a pure
placeholder and a root for the low-bandwidth allocations
mainly done by vmem metadata.  That donates until module
unload a physical-memory-dependent fixed allocation to the
"bucket_heap" arena, which serves as an analogue of
Illumos's lowest-level vmem arena.

The bucket_heap exchanges memory dynamically with a set of
bucket arenas which are organized by powers-of-two, with
each bucket taking bucket_heap allocation requests at that
power-of-two at maximum, and 1 + the next smaller
power-of-two at minimum.

These bucket arenas in turn exchange memory with xnu in
chunks that are by default the larger of 1 megabyte or
four times the maximum power-of-two.  The goal is to
reduce internal fragmentation and to dramatically reduce
the amount of memory that is exchanged with xnu's
alloator, which can be slow, and which copes so badly with
high frequency alloc/free that the entire operating system
can halt on a watchdog condition.   Segregation into
buckets is a good trade-off between reduction of waste due
to fragmentation and the ability to quickly and
dynamically reduce memory footprint in response to memory
demands elsewhere in the system (notably in userland).

A set of allocate and free functions sit between the
bucket_heap and the bucket arenas, and another set between
the bucket arenas and XNU.   The goal in each is to
minimize the calls to XNU's kernel memory allocator, and
to spread out unavoidable allocations and frees so that
XNU has time to process the requests.

Additionaly, the bucket arenas' allocation functions will
avoid making unconditional allocations.  A conditional
allocation is only made when XNU's vm_page_free_count is
above a threshold.   Pressure is signalled when a
conditional allocation fails, causing ARC to halt its
growth for at least a minute, and try to reduce its size.
Most failed conditional allocations will be re-tried for a
period of time, and some may be promoted into an
uncontidional allocation.    The goal is to avoid
user-visible effects such as a stall in teh zfs subsystem
or an increase in memory pressure.

Log entries indicating pressure, waiting threads, activity
by arc_reclaim_thread, or "arc growtime expired" are for
diagnostic purposes and are harmless and may be ignored.

As noted, the set-up and tear-down of the vmem system
differs dramatically between Illumos and our
runtime-unloadable kernel extension.  Many diagnostic
messages are logged when the module is loaded or unloaded;
they are harmless and may be ignored.

Finally, in SPL and ZFS we maintain separate arenas for zfs
file data and zfs metadata, which gives us quantum caching
gains, reduction in internal fragmentation, and visibility
to the spl_free generating thread and to developers and
interested sysadmins.  These arenas are clients of the
bucket_heap and in turn service kmem_cache_alloc and
similar requests related to ARC, zio and other subsystems
which process the relevant data.  Other kmem clients are
ultimately rooted to kmem_va, which is also a client of
the bucket_heap.