Use the right arenas.
The special alloc & free functions are called with vmp ==
spl_root_arena_parent, which is a placeholder, that will
never have memory in it.  Instead, use the appropriate
caller arenas, whether that's spl_heap_arena
("bucket_heap" in kstats) or a bucket arena.

Two D.R.Y. functions relating bucket with size.

Kstats on free memory discovery in the calling or called arenas.