stretch vmem_bucket_allocate for loop
As there is at least one waiter in the loop, it-or-they
should stay in vba() longer for two reasons: firstly,
another thread is spending a lot of time in vmem_alloc(),
which means it is likely waiting on xnu for memory, so
this thread would be too; secondly, we can continue
probing the bucket_arena (the parent) to see if another
thread has produced free memory there, which we cannot use
once we have left the for loop.

Even though the iter test is longer, the most common exit
path will be when waiters drops to 1, since most calls to
vmem_alloc() will not proceed to vmem_xalloc(), as can be
seen by comparing each bucket arena's alloc vs
parent_alloc kstats.