adjust vba() based on new kstat counts
Set buckets_busy_allocating bit even if the for loop is not
entered, or is exited because waiters drops to 1, thus
preventing subsequent threads acting on the same
vmem_canalloc_semiatomic() information on their first
pass through the for loop.

Use an atomic test-and-set on buckets_busy_allocating
within the for loop to avoid the rare case where the
non-atomic test and set could let two threads out of the
for loop (especially with the |= moved to outside the for loop).

Set a longer wait time in the cv_timedwait_hires for
non-VM_PUSHPAGE sleepable allocations, which influences
the kernel waitq.

Move and update comments.