Deal with memory shortfalls in vba()
If on entry to vba() for a VM_SLEEP allocation the thread
discovers that there is no memory in the target bucket
then it should raise pressure and then wait at least
briefly for the pressure to work before proeeding.

Although xat() also does this, there is an important
difference: for VM_SLEEP allocations, vba() can take
advantage of pressure causing memory to be freed in the
bucket heap (notably, in the initial fixed allocation)
while xat() cannot.