reduce CPU: do IOSleep rather than IODelay for long delay()s
IODelay(microseconds) does a timed spin and keeps the CPU busy.

Most delay(ticks) calls are for several ticks (a tick is
10 milliseconds), and this poses a particularly
counterproductive problem when (for example)
kstat.zfs.darwin.tunable.zfs_scrub_delay is set high in an
attempt to reduce the CPU impact of a fast scrub.

Instead, call IOSleep(milliseconds).   IOSleep may not
return exactly on schedule; if we wake up early, call
IOSleep with the residual sleep time.

Reference:

https://developer.apple.com/library/content/documentation/Darwin/Conceptual/KernelProgramming/services/services.html