use IOSleep in xat_bail, make some timeouts shorter
IOSleep(1) instead of kpreempt(PREEMPT_SYNC) where we are
OK with a long suspension, in xat_bail and in vba.

Half the amount of time waiting in xat_bail().

When a timeout exit is blocked in vba, we don't want to wait
another 30-60 msec before trying to escape the for loop again.