respect higher priority allocations in vba()
Higher priority allocations are those that are:

i) those that do not wait in the first place:
VM_NOSLEEP, VM_ABORT and/or VM_PANIC flags

ii) flagged higher prority, VM_SLEEP or not:
VM_PUSHPAGE

If a thread sees memory available but there is a higher
priority thread in or below vmem_bucket_alloc() then
the thread skips testing the buckets_busy_allocating array
and stays in the for loop.

Additionally, properly batch-count the blocked timeouts
as *blocked* timeouts.