vba: set timeout earlier for PUSHPAGE + SLEEP allocs
Problem:

A non-PUSHPAGE allocation in the vmem_bucket_alloc() for
loop is cannot exit after a PUSHPAGE allocation enters; a
spatter of PUSHPAGE allocations in relatively low memory
can stall out the non-PUSHPAGE allocation for a long time.

1. PUSHPAGE enters, locking out non-PUSHPAGE; PUSHPAGE
   goes to sleep.
2. The PUSHPAGE sleep is interrupted (cv_timedwait_hires returns 0)
   so timeout is not set, memory is still not available,
   and so cv_timedwait_hires is entered again.
3. repeat step 2 until ~290 msec elapses and the PUSHPAGE finally
   sees timeout set to 2.   Meanwhile, non-PUSHPAGE is still
   stuck in the for loop.
4. PUSHPAGE exits for loop and enters vmem_alloc();
   meanwhile another PUSHPAGE enters the for loop, go to
   step 2.

This pattern was observed during a scrub.

Solution:

If doing a PUSHPAGE allocation, set timeout flag after 40
ms if it has not yet been set (by cv_timedwait_hires
returning -1) and if there are other waiters.