VMC_TIMEFREE : keep freelist sorted by span creation time
VMC_TIMEFREE is a vmem arena create time flag.

When set on an arena:

Save gethrtime() whenever a span is imported.

When a segment is created from the span, the segment
inherits the span import time.

When adding a segment to the freelist via
vmem_freelist_insert(), call a specialized function,
vmem_freelist_insert_sort_by_time, which does a
bounded-linear search (max_walk_steps) of the freelist for
a segment younger than the segment to be inserted.   The
segment is inserted before the younger segment.

This keeps the freelists for the arena sorted by span
create time.

The oldest spans are the ones being held "hostage" by old
allocations that are most likely to be the hottest ARC
buffers or the hottest heap allocations, so we want them
to be allocated from first.

Keeping the oldest spans active in this way will reduce
fragmentation compared to simply pushing every segment onto
head of the appropriate freelists.

i.e., for the bucket areans, using the oldest spans first
is more likely to cause spans to be freed back to xnu than
using the oldest spans last and hoping that ARC or the
heap releases the hostages.

Additionally, added VERIFY()s in vmem_init(), sprinkled some ASSERT()s,
and some dprintfs -- new C99-style "dprintf(...)" macro,
and commented out the IOSleep() that would make the printf
output more likely to be visible in a huge rush of vmem_free()s.