set VM_BESTFIT if VMC_TIMEFREE instant-fit alloc fails
The resulting scan of the lower freelist is offset by the
chance that a (more expensive) trip to vmem_bucket_alloc()
and ultimately the xnu allocator will be avoided because a
large-enough segment is found on next-lower freelist.

In the worst case, the next-lower freelist is long but
doesn't have a large enough free segment on it.   We would
have gone to vba() anyway, but now we've walked a long
list.

However, the next-lower freelist is likely to be fairly
short and because of bucketing is also likely to have a
large enough free segment on it if it has any segments at
all.    Walking the next-lower freelist is likely to take
on the order of microseconds; a trip to vba is many milliseconds.

(The cheapest case is that the next-lower freelist is of
zero length, in which case we've added mere nanoseconds;
the trip down to xnu's allocator will take many milliseconds.)