fix ks_set_cp's bad || check and harden the maybe-panic case
oops,
"if (ks->cp_b == cp || ks->cp_b == cp)"
should have tested ks->cp_a for == cp as well.

The compiler ought to have have warned about the duplicate
testing, but on the other hand cp_b is _Atomic so could in
principle have changed during the test (although the
assembly does not actually allow that).

Additionally, check for ks->cp_{a,b} != cp before entering
the "ERROR" condition, since the race may set one of
those to cp (and thus not NULL).

The panic must be the result of a peculiar race, but
unfortunately we don't know the state of ks->cp_a or
ks->cp_b, so break apart the long if (|| || || ...)
before the panic.

Also, harden against an unlikely but possible NULL deref.

2017-01-17 TZ=UTC
[05:36:41] <lundman> woah panic
[05:37:03] <lundman> rotte: this could be yours
[05:38:22] <lundman> never seen it before
[05:38:31] <lundman> panic(cpu 1 caller 0xffffff7f8666e0ac): "iskvec[255] : cannot set for zio_data_buf_131072"@spl-kmem.c:6411
[05:38:39] <lundman> 0xffffff909f773690 : 0xffffff7f8666e0ac net.lundman.spl : _ks_set_cp + 0x14c
[05:38:42] <lundman> 0xffffff909f7736d0 : 0xffffff7f8666b0a6 net.lundman.spl : _spl_zio_set_no_grow + 0x26
[05:38:45] <lundman> 0xffffff909f7736f0 : 0xffffff7f8666b050 net.lundman.spl : _spl_zio_kmem_cache_alloc + 0x80

squash! harden ks_set_cp against race that caused a panic