new function: spl_arc_reclaim_needed(bytes, zp)
This function is called from arc_adapt() which previously
would have called arc_reclaim_needed(), which is true when
spl_free < 0.   Since spl_free is likely to be negative
for many milliseconds, many unnecessary cv_signals are
sent to wake up arc_reclaim_thread when there is a
streaming reader or writer (or both), and those wakeups
are between useless (arc_reclaim_thread is already
running) or of very low value (the wakeup happens almost
immedaitely after a reclaim has run when there is little
that can be evicted).

zp is a pointer to zio_buf_cache or zio_data_buf_cache,
both of which are defined and populated in zio.c and thus
are not available for linking in spl.kext

Because of qcaching, checking the buckets for free memory
would give a false memory-available signal, with the
practical result of increased bucket fragmentation.

Instead we check to see if the kmem_cache defined by the
argument is holding free spans; if it is, we return
B_FALSE and avoid a reclaim.

If useful memory is tight (low_threshold), return B_TRUE,
but if there is lots of useful memory, return B_FALSE.

Otherwise, if spl_free < 0, return B_TRUE, as
arc_reclaim_needed would.

Add a new kstat (spl_arc_reclaim_avoided) for monitoring.

fine-tune spl_arc_reclaim_needed()

if useful memory is tight, return B_TRUE