kmem: sort partial slabs by hires time of slab creation
When two slabs have the same partial weight (i.e., they
are approximately as full as one another), then Illumos
breaks the tie by examining the addresses of the slabs.

Since the addresses of the slabs are not very meaningful
in our implementation, first try to break the tie using
the slab age, preferring the older slab.