kmem_va above 8k is negligible, so reduce its qcaching size
This moves kmem_alloc allocations for slabs 8k and below
into the 32k bucket.   Above kmem_alloc_8192 only
kmem_alloc_12288.slab_size is 16k or lower, but those
allocations are rare (~thousands compared to tens- or
hundreds of millions for 2k and sub 512-byte allocs).

Moving kmem_va out of the bucket_65536 is likely to help
reduce the "kidnapping-by-sharing" of the bucket_65536
slabs, wherein a qcached kmem_va segment is mixed in with
many qcached zfs_qcache segments.