anti-fragmentation logic for abd's benefit
1. New flag, KMC_ARENA_SLAB for explicit minimum-import arenas

When set at kmem_cache_create time, use half the parent
arena's minimum import size instead of its quantum
to size the kmem cache's slab size.

ABD chunks are 1k so the kmem_cache_create() logic chooses
a PAGESIZE kmem cache slab size.

Increasing the kmem cache slab size is good because the
kmem_cache sorting logic (using avl trees and a hashing
distance that sorts finely at the extremes) is faster and
generally superior to VMC_TIMEFREE's simpler sorting logic
especially since the latter is linear with the number of
items on the (arena) freelists, and having a
larger (cache) slab size means fewer items on the
freelist.

KMC_ARENA_SLAB dramatically reduces the typical amount of
partial (arena) slabs in the abd_chunk arena and thereby
reduces kstat.spl.misc.spl_misc.os_mem_alloc for a given
arc size too.   (It also marginally improves performance
as the number of cpus increases).

2. Expose kmem_depot_ws_zero(), so it can be called when
there has been successful abd move, as moving tends to
create large numbers of full magazines which would linger
in the depot layer as an unnecessarily inflated working
set.   The working set will grow as usual if the abd cache
is busy between the _zero() call and the next reap, so
this mainly affects moves that release memory while the
system is quiescent.

3. Introduce kmem_cache_free_to_slab, sysctl-tunable, default off.

It defaults off because it increases kmem_free() latency
both by running extra code (rather than updating a
non-full magazine in the typical case) and lock contention
on the arena mutex.   It rapidly reduces fragmentation in
the trees rooted at the heap and the zfs_qcache arenas,
which under some loads (lots of metadata updating, in
particular) can become highly fragmented.

When on, kmem_free() calls (to caches whose vmem arenas
can maintain excess memeory) will test whether there is
apparent fragmentation in the cache's parent arena and in
the system generally.  When there is such fragmentation,
frees will go the the slab layer rather than the depot
layer.

We keep the fast path mechanism of freeing to the loaded
cpu magazine until the magazine is full.

If the loaded magazine is full, we swap loaded with last-loaded
magazine as usual if there is no fragmentation in the
parent arena.

Howeer, if the loaded magazine is full and the parent
arena is fragmented, we return the memory to the slab
layer.

This a shortcut of the (working-set dependent) release of
full magazines from the depot layer at reap time.

The thresholds are sufficiently high that small, dense, or
busy kmem_caches will rarely take this path.