if fragmented, return excess kmem_cache_free() to the slab layer
Deal with multiple back-to-back frees hitting the same
cpu (which generally means back-to-back frees in the same
thread, which when scattered ABDs are in use can certainly
happen during ARC evictions and ABD moves and a handful of
other cases).

This is a deliberate trade-off between the lowest latency of
kmem_cache_free() and being able to return memory from a
kmem_cache during "big free event", favouring the latter.

We keep the fast path mechanism of freeing to the loaded
cpu magazine until the magazine is full.

If the loaded magazine is full, we swap loaded with last-loaded
magazine as usual if there is no fragmentation in the
parent arena.

Howeer, if the loaded magazine is full and the parent
arena is fragmented, we return the memory to the slab
layer.

This a shortcut of the (working-set dependent) release of
full magazines from the depot layer at reap time.

The thresholds are sufficiently high that small, dense, or
busy kmem_caches will rarely take this path.