EXPERIMENT: taskq should use physical_ncpus
when configured to use a percentage of cpus, taskq uses
max_ncpus == 100%.  max_ncpus is the limit of the number
of logical cpus the system has (at present, this is always
also the number of logical cpus that are active).

instead, use the number of physical cpus.

(it is probably OK to use hw.activecpu, which on intel
macintoys will also be the same as logicalcpu_max, so if
performance is poor with this commit, that will be the
fix).