switch to using IOMemAllocAligned in old spl
import several changes from new o3x

in particular stop using kernel_memory_allocate in favour
of IOMallocAlign.  The vmem layer above requires natural
alignment, so we use a bit twiddling device to round up to
the enclosing power of two (minimum PAGESIZE) for
the (rare) cases where osif_alloc is asked for a
non-power-of-two or a small allocation.

We also change the alignment of the BSS block of initial
memory to be ARM-friendly; this originally motivated the
change to IOMallocAlign, but that turns out to have been
a good move to stop using an unsupported MACH-level interface.

IOFree has logic underneath that tracks the largest
allocations seen via IOMalloc/IOMallocAligned and will
panic if it sees anything larger.    Switch the initial
populating vmem_alloc to vmem_xalloc, which fixes the
panic.  Vmem_alloc did extra unnecessary segment
bookkeeping and would break up the initial allocation into
several, or even many, smaller ones.   The vmem_fini call
to vmem_xfree would just return the entire chunk at once.

The bucket arenas should be adapted to larger quanta in
the face of having to supply an alignment size to
IOMallocAligned().  This simplifies bucket arena setup
while retaining good performance.